renamed style.py to convenience.py for clarity
