+++ convenience.py
