+++ interpolation.py
@@ -6,7 +6,7 @@ from helper import *
 from config.GlobalVariables import *
 from SynthesisNetwork import SynthesisNetwork
 from DataLoader import DataLoader
-import style
+import convenience
 
 L = 256
 
@@ -35,19 +35,19 @@ def main(params):
     if params.output == "image" and params.interpolate == "writer":
         if len(params.blend_weights) != len(params.writer_ids):
             raise ValueError("writer_ids must be same length as writer_weights")
-        im = style.sample_blended_writers(params.blend_weights, params.target_word, net, all_loaded_data, device)
+        im = convenience.sample_blended_writers(params.blend_weights, params.target_word, net, all_loaded_data, device)
         im.convert("RGB").save(f'results/blend_{"+".join([str(i) for i in params.writer_ids])}.png')
     elif params.output == "grid" and params.interpolate == "character":
         if len(params.grid_chars) != 4:
             raise ValueError("grid_chars must be given exactly four characters")
-        im = style.sample_character_grid(params.grid_chars, params.grid_size, net, all_loaded_data, device)
+        im = convenience.sample_character_grid(params.grid_chars, params.grid_size, net, all_loaded_data, device)
         im.convert("RGB").save(f'results/grid_{"+".join(params.grid_chars)}.png')
     elif params.output == "video" and params.interpolate == "writer":
-        style.writer_interpolation_video(params.target_word, params.frames_per_step, net, all_loaded_data, device)
+        convenience.writer_interpolation_video(params.target_word, params.frames_per_step, net, all_loaded_data, device)
     elif params.output == "video" and params.interpolate == "character":
-        style.char_interpolation_video(params.blend_chars, params.frames_per_step, net, all_loaded_data, device)
+        convenience.char_interpolation_video(params.blend_chars, params.frames_per_step, net, all_loaded_data, device)
     elif params.interpolate == "randomness":
-        style.mdn_video(params.target_word, params.num_mdn_samples, params.scale_randomness, params.max_randomness, net, all_loaded_data, device)
+        convenience.mdn_video(params.target_word, params.num_mdn_samples, params.scale_randomness, params.max_randomness, net, all_loaded_data, device)
     else:
         raise ValueError("Invalid task")
